+++ NEWS.md
@@ -1,6 +1,6 @@
 # Version 0.1.1
 * Added the geographic distance penalty for selecting the nearest predicted neighbours (`xy_penalty` and `xy_stats` parameters)
-* Added an option for excluding self-assessment for benchmark sites.
+* Added an option for excluding self-assessment for benchmark sample sites (default).
 
 # Version 0.1.0
 * Translated all legacy HCAS codes to R and Rcpp